chore: remove the previous implementation of DeferredPromise
