+++ src/DeferredPromise.ts
@@ -1,173 +1,33 @@
 import {
+  createDeferredExecutor,
+  DeferredPromiseExecutor,
   PromiseState,
-  RejectFunction,
   ResolveFunction,
+  RejectFunction,
 } from './createDeferredExecutor'
 
-export type ExtractItemType<A extends Array<any>> = A extends Array<infer T>
-  ? T
-  : never
-
 /**
- * Represents the completion of an asynchronous operation.
- * Can be resolved or rejected outside of its closure.
- *
  * @example
- * const numberReady = new DeferredPromise()
- * numberReady.resolve(123)
- *
- * @example
- * const portReady = new DeferredPromise()
- * portReady.reject(new Error('Port is already in use'))
+ * function getPort() {
+ *   const portReadyPromise = new DeferredPromise<number>()
+ *   port.on('ready', (port) => portReadyPromise.resolve(port))
+ *   port.on('error', (error) => portReadyPromise.reject(error))
+ *   returtn portReadyPromise
+ * }
  */
-export class DeferredPromise<Data extends any = void> {
+export class DeferredPromise<Data = void> {
+  #promise: Promise<Data>
+  #executor: DeferredPromiseExecutor<Data>
+
   public resolve: ResolveFunction<Data>
   public reject: RejectFunction
-  public state: PromiseState
-  public result?: Promise<Data>
-  public rejectionReason?: unknown
-
-  private thenCallbacks: Array<ResolveFunction<unknown> | undefined>
-  private catchCallbacks: Array<RejectFunction<unknown>>
-  private promise: Promise<unknown>
 
   constructor() {
-    this.thenCallbacks = []
-    this.catchCallbacks = []
+    this.#executor = createDeferredExecutor<Data>()
+    this.#promise = new Promise<Data>(this.#executor)
 
-    this.promise = new Promise<Data>((resolve, reject) => {
-      this.resolve = (data) => {
-        if (this.state !== 'pending') {
-          return
-        }
-
-        this.state = 'fulfilled'
-        this.result = this.waterfall(data, this.thenCallbacks) as Promise<Data>
-        this.result.then((result) => {
-          resolve(result)
-        })
-      }
-
-      this.reject = (reason) => {
-        if (this.state !== 'pending') {
-          return
-        }
-
-        console.trace('reject', reason)
-
-        this.state = 'rejected'
-        this.rejectionReason = reason
-        reject(reason)
-
-        // this.waterfall(reason, this.catchCallbacks).then((result) => {
-        //   console.warn('reject waterfall result:', result)
-
-        // })
-      }
-    }).catch((reason) => {
-      console.log('REJE')
-      return this.waterfall(reason, this.catchCallbacks)
-    })
-
-    this.state = 'pending'
-    this.result = undefined
-    this.rejectionReason = undefined
-  }
-
-  // /**
-  //  * Creates a Promise that reduces all the `.then()` callbacks
-  //  * over the result that the deferred promise has resolved with.
-  //  * Returns the result of that.
-  //  */
-  // private async toResultPromise(data: Data): Promise<Data> {
-  //   let result: Data = data
-  //   let tuple: ExtractItemType<typeof this.transforms>
-
-  //   console.log('--> toResultPromise', this.transforms)
-
-  //   while ((tuple = this.transforms.shift())) {
-  //     const [next, reject] = tuple
-
-  //     try {
-  //       result = (await next(result)) as Data
-  //     } catch (error) {
-  //       reject?.(error)
-  //     }
-  //   }
-
-  //   return result
-  // }
-
-  private async waterfall(
-    input: unknown,
-    fns: Array<(...args: unknown[]) => unknown>
-  ): Promise<unknown> {
-    let result = input
-    let fn: (...args: unknown[]) => unknown
-
-    while ((fn = fns.shift())) {
-      try {
-        result = await fn(result)
-      } catch (error) {
-        console.trace('HEY THIS IS WRONG', error)
-      }
-    }
-
-    return result
-  }
-
-  /**
-   * Attaches callbacks for the fulfillment and rejection of the Promise.
-   */
-  public then<ResolveData = Data, RejectionReason = never>(
-    onFulfilled?: ResolveFunction<Data>,
-    onRejected?: RejectFunction<RejectionReason>
-  ): DeferredPromise<ResolveData | RejectionReason> {
-    const chainable = new DeferredPromise<ResolveData | RejectionReason>()
-
-    /**
-     * @todo Delegate all transforms to the right-most (current) Promise.
-     * Since `.resolve()` will be called only on that last Promise,
-     * other preceding promises won't be able to yield value.
-     */
-    if (typeof onFulfilled === 'function') {
-      chainable.thenCallbacks.push(...this.thenCallbacks, onFulfilled)
-    }
-
-    if (typeof onRejected === 'function') {
-      chainable.catchCallbacks.push(...this.catchCallbacks, onRejected)
-    }
-
-    this.promise.then(
-      (result: ResolveData | RejectionReason) => {
-        chainable.resolve(result)
-      },
-      (error) => {
-        console.error('rejected!', error)
-        chainable.reject(error)
-      }
-    )
-
-    return chainable
-  }
-
-  /**
-   * Attaches a callback for only the rejection of the Promise.
-   */
-  public catch<RejectReason = never>(
-    onRejected?: RejectFunction<RejectReason>
-  ): DeferredPromise<Data | RejectReason> {
-    return this.then(null, onRejected)
-  }
-
-  /**
-   * Attaches a callback that is invoked when
-   * the Promise is settled (fulfilled or rejected). The resolved
-   * value cannot be modified from the callback.
-   */
-  public finally(onFinally?: () => void): DeferredPromise<Data> {
-    // this.promise = this.promise.finally(onFinally)
-    return this
+    this.resolve = this.#executor.resolve
+    this.reject = this.#executor.reject
   }
 
   static get [Symbol.species]() {
@@ -176,5 +36,130 @@ export class DeferredPromise<Data extends any = void> {
 
   get [Symbol.toStringTag]() {
     return 'DeferredPromise'
+  }
+
+  public get state(): PromiseState {
+    return this.#executor.state
+  }
+
+  public get rejectionReason(): unknown {
+    return this.#executor.rejectionReason
+  }
+
+  public then<FulfillmentResult = Data, RejectionResult = never>(
+    onFulfilled?: ResolveFunction<Data, FulfillmentResult> | null,
+    onRejected?: RejectFunction<RejectionResult> | null
+  ): DeferredPromise<FulfillmentResult | RejectionResult> {
+    const derivedPromise = new DeferredPromise<
+      FulfillmentResult | RejectionResult
+    >()
+
+    const resolveDerivedPromise = (result: any): void => {
+      if (typeof onFulfilled === 'function') {
+        try {
+          const nextResult = onFulfilled(result)
+          derivedPromise.#executor.resolve(nextResult as any)
+        } catch (error) {
+          rejectDerivedPromise(error)
+        }
+
+        return
+      }
+
+      /**
+       * @note Use the executor directly because the "resolve" method
+       * always gets overridden to point to the previous Promise in the chain.
+       */
+      derivedPromise.#executor.resolve(result)
+    }
+
+    const rejectDerivedPromise = (result: RejectionResult): void => {
+      if (typeof onRejected === 'function') {
+        try {
+          const nextReason = onRejected(result)
+
+          /**
+           * @note If the rejection has a handler callback ("catch"),
+           * resolve the chained promise instead of rejecting it.
+           * The "catch" callback is there precisely to prevent
+           * uncaught promise rejection errors.
+           */
+          derivedPromise.#executor.resolve(nextReason as any)
+
+          /**
+           * @note Since the handled rejected promise was resolved internally,
+           * still mark its state as rejected. Do so in the next microtask
+           * because executor rejects the promise in the next task as well
+           */
+          queueMicrotask(() => {
+            derivedPromise.#executor.state = 'rejected'
+            derivedPromise.#executor.rejectionReason = nextReason
+          })
+        } catch (error) {
+          rejectDerivedPromise(error)
+        }
+
+        return
+      }
+
+      derivedPromise.#executor.reject(result)
+    }
+
+    switch (this.state) {
+      case 'pending': {
+        /**
+         * @note Override the chainable "resolve" so that when called
+         * it executes the resolve of the Promise from which it was
+         * chained from.
+         *
+         * @example
+         * const p1 = new DeferredPromise().then(A).then(B)
+         * p1.resolve()
+         *
+         * Here, "p1" points to the rightmost promise (B) but the
+         * order of resolving the value must be left-to-right:
+         * Initial -> A -> B
+         */
+        derivedPromise.resolve = this.resolve as ResolveFunction<
+          FulfillmentResult | RejectionResult
+        >
+        derivedPromise.reject = this.reject
+
+        this.#promise.then(resolveDerivedPromise, rejectDerivedPromise)
+
+        break
+      }
+
+      case 'fulfilled': {
+        resolveDerivedPromise(this.#executor.result)
+        break
+      }
+
+      case 'rejected': {
+        rejectDerivedPromise(this.#executor.rejectionReason as RejectionResult)
+        break
+      }
+    }
+
+    return derivedPromise
+  }
+
+  public catch<Result>(
+    onRejected?: RejectFunction<Result>
+  ): DeferredPromise<Data | Result> {
+    return this.then(undefined, onRejected)
+  }
+
+  public finally(onSettled?: (() => void) | null) {
+    return this.then(
+      (result) => {
+        onSettled?.()
+        return result
+      },
+      (reason) => {
+        onSettled?.()
+        throw reason
+      }
+    )
   }
 }
