+++ test/DeferredPromise/reject.test.ts
@@ -1,4 +1,4 @@
-import { DeferredPromise } from '../../src/_DeferredPromise'
+import { DeferredPromise } from '../../src/DeferredPromise'
 
 it('can be rejected without any reason', async () => {
   const promise = new DeferredPromise<void>().catch(() => {})
